requirements.txt

streamlit
sentence-transformers
faiss-cpu
PyPDF2
pyttsx3
torch
